Refactor YTMYHCHResponse and TwitchHandler logic
- Changed `Views` property type to `long` in `YTMYHCHResponse` for larger counts.
- Updated `RequestObject` initialization to use C# 9.0 syntax.
- Simplified `videoId` assignment logic in `TwitchHandler`.
- Set `Artist` to an empty string and `Title` to full title in `RequestObject`.
- Adjusted queue check logic to return a message if the song is already queued.
- Used new list initialization syntax in `GetQueueItems`.
- Updated `TwitchRequestUser` creation to use new object initialization syntax.

+++ Songify Slim/Models/YTMD/YTMYHCHResponse.cs
@@ -11,7 +11,7 @@ namespace Songify_Slim.Models.YTMD
             public string Title { get; set; }
             public string AlternativeTitle { get; set; }
             public string Artist { get; set; }
-            public int Views { get; set; }
+            public long Views { get; set; }
             public DateTime UploadDate { get; set; }
             public string ImageSrc { get; set; }
             public bool IsPaused { get; set; }
